+++ ValRougelike/Common/ValConfig.cs
@@ -11,6 +11,7 @@ public class ValConfig
     public static ConfigEntry<string> ItemsNotSkillChecked;
     public static ConfigEntry<float> DeathSkillPerLevelBonus;
     public static ConfigEntry<int> MinimumEquipmentRetainedOnDeath;
+    public static ConfigEntry<int> MaximumEquipmentRetainedOnDeath;
     public static ConfigEntry<float> MaxPercentResourcesRetainedOnDeath;
     public static ConfigEntry<float> MaxPercentTotalItemsRetainedOnDeath;
     public static ConfigEntry<float> GainedSkillLossFactor;
@@ -23,6 +24,7 @@ public class ValConfig
     public static ConfigEntry<bool> FoodLossOnDeath;
     public static ConfigEntry<bool> FoodLossOnDeathBySkillLevel;
     public static ConfigEntry<bool> ShowDeathMapMarker;
+    public static ConfigEntry<bool> ItemsSavedToTombstone;
     //public static ConfigEntry<bool> EffectRemovalOnDeath;
 
     public static ConfigEntry<float> SkillProgressUpdateCheckInterval;
@@ -41,10 +43,12 @@ public class ValConfig
     {
         DeathSkillPerLevelBonus = BindServerConfig("DeathProgression","DeathSkillPerLevelBonus",1f,"How impactful death skill progression is. This impacts how much each level improves your skill and item retention.", false, 0f, 10f);
         MinimumEquipmentRetainedOnDeath = BindServerConfig("DeathProgression","MinimumEquipmentRetainedOnDeath",2,"The minimum amount of Equipment that can be retained on death, depends on players individual skill.", true, 0, 30);
+        MaximumEquipmentRetainedOnDeath = BindServerConfig("DeathProgression", "MaximumEquipmentRetainedOnDeath", 10, "The maximum amount of Equipment that can be retained on death, depends on players individual skill.", true, 0, 30);
         MaxPercentResourcesRetainedOnDeath = BindServerConfig("DeathProgression","MaxPercentResourcesRetainedOnDeath",20f,"The maximum amount of Resources that can be retained on death, depends on players individual skill.", true, 0f, 100f);
         MaxPercentTotalItemsRetainedOnDeath = BindServerConfig("DeathProgression","MaxPercentTotalItemsRetainedOnDeath",90f,"The maximum amount of total items that can be retained on death, depends on players individual skill.", true, 0f, 100f);
         OnlyXPLossFromSkillGains = BindServerConfig("SkillLossModifiers", "OnlyXPLossFromSkillGains", true, "When enabled, you can only loose XP gained since the last death. Repeated deaths regardless of time without skill gains will not result in XP loss.");
         GainedSkillLossFactor = BindServerConfig("SkillLossModifiers", "GainedSkillLossFactor", 0.2f, "The percentage of skills that are lost when dying.", false, 0f, 1f);
+        ItemsSavedToTombstone = BindServerConfig("DeathProgression", "ItemsSavedToTombstone", false, "Items are saved to your tombstone instead of saved to your character.");
 
         ItemsNotSkillChecked = BindServerConfig("DeathProgression", "ItemsNotSkillChecked", "Tin,TinOre,Copper,CopperOre,CopperScrap,Bronze,Iron,IronScrap,Silver,SilverOre,DragonEgg,chest_hildir1,chest_hildir2,chest_hildir3,BlackMetal,BlackMetalScrap,DvergrNeedle,MechanicalSpring,FlametalNew,FlametalOreNew", "List of items that are not rolled to be saved through death progression.");
         ItemsNotSkillCheckedAction = BindServerConfig("DeathProgression", "ItemsNotSkillCheckedAction", "dropOnDeath",